# docs/api/openapi/patients-providers.yaml
openapi: 3.1.1
info:
  title: MedFlow Patients and Providers Fragment
  version: 0.1.0
paths:
...
  /api/v1/patients/{patientId}/contacts:
    get:
      operationId: ListPatientContactsQuery
      tags: [Patients]
      summary: List patient contacts
      description: Returns tenant-scoped patient contacts ordered by primary flag, emergency flag, name, and creation time.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientIdPath'
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Patient contacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientContactCollectionResponse'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Missing tenant scope or patient read permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Patient not found in the active tenant scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    post:
      operationId: CreatePatientContactCommand
      tags: [Patients]
      summary: Create a patient contact
      description: Creates a tenant-scoped patient contact. At least one of `phone` or `email` is required.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientIdPath'
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientContactRequest'
      responses:
        '201':
          description: Patient contact created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientContactMutationResponse'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Missing tenant scope or patient write permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Patient not found in the active tenant scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '422':
          description: Invalid contact payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/patients/{patientId}/contacts/{contactId}:
    patch:
      operationId: UpdatePatientContactCommand
      tags: [Patients]
      summary: Update a patient contact
      description: Updates a tenant-scoped patient contact while preserving the single-primary-contact invariant.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientIdPath'
        - $ref: '#/components/parameters/PatientContactIdPath'
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientContactRequest'
      responses:
        '200':
          description: Patient contact updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientContactMutationResponse'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Missing tenant scope or patient write permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Patient or contact not found in the active tenant scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '422':
          description: Invalid contact payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    delete:
      operationId: DeletePatientContactCommand
      tags: [Patients]
      summary: Delete a patient contact
      description: Hard deletes a tenant-scoped patient contact and records a patient audit event.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientIdPath'
        - $ref: '#/components/parameters/PatientContactIdPath'
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Patient contact deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientContactMutationResponse'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Missing tenant scope or patient write permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Patient or contact not found in the active tenant scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /api/v1/patients/{patientId}/documents:
    get:
      operationId: ListPatientDocumentsQuery
      tags: [Patients]
      summary: List patient document metadata
      description: Returns patient document metadata newest first. Responses never include public file URLs or storage-internal paths.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientIdPath'
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Patient document metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDocumentCollectionResponse'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Missing tenant scope or patient read permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Patient not found in the active tenant scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    post:
      operationId: UploadPatientDocumentCommand
      tags: [Patients]
      summary: Upload a patient document
      description: >
        Uploads a patient document to the private attachments disk and returns metadata only.
        If `title` is omitted it defaults to the uploaded filename.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientIdPath'
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadPatientDocumentRequest'
      responses:
        '201':
          description: Patient document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDocumentMutationResponse'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Missing tenant scope or patient write permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Patient not found in the active tenant scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '422':
          description: Invalid document upload payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/patients/{patientId}/documents/{docId}:
    get:
      operationId: GetPatientDocumentQuery
      tags: [Patients]
      summary: Get patient document metadata
      description: Returns document metadata only for one patient document in tenant scope.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientIdPath'
        - $ref: '#/components/parameters/PatientDocumentIdPath'
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Patient document metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDocumentResponse'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Missing tenant scope or patient read permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Patient or document not found in the active tenant scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    delete:
      operationId: DeletePatientDocumentCommand
      tags: [Patients]
      summary: Delete a patient document
      description: Deletes patient document metadata and performs best-effort deletion of the private attachment file.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientIdPath'
        - $ref: '#/components/parameters/PatientDocumentIdPath'
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Patient document deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDocumentMutationResponse'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Missing tenant scope or patient write permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Patient or document not found in the active tenant scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PatientIdPath:
      name: patientId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PatientContactIdPath:
      name: contactId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PatientDocumentIdPath:
      name: docId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TenantIdHeader:
      name: X-Tenant-Id
      in: header
      required: true
      description: Active tenant scope for the request.
      schema:
        type: string
        format: uuid
...
  schemas:
...
    PatientContact:
      type: object
      required:
        - id
        - patient_id
        - name
        - is_primary
        - is_emergency
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        name:
          type: string
        relationship:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        is_primary:
          type: boolean
        is_emergency:
          type: boolean
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreatePatientContactRequest:
      type: object
      required: [name]
      description: At least one of `phone` or `email` is required.
      properties:
        name:
          type: string
          maxLength: 160
        relationship:
          type: string
          nullable: true
          maxLength: 120
        phone:
          type: string
          nullable: true
          maxLength: 32
        email:
          type: string
          format: email
          nullable: true
          maxLength: 255
        is_primary:
          type: boolean
        is_emergency:
          type: boolean
        notes:
          type: string
          nullable: true
          maxLength: 5000
    UpdatePatientContactRequest:
      type: object
      description: At least one field is required. The resulting contact must still contain `phone` or `email`.
      properties:
        name:
          type: string
          maxLength: 160
        relationship:
          type: string
          nullable: true
          maxLength: 120
        phone:
          type: string
          nullable: true
          maxLength: 32
        email:
          type: string
          format: email
          nullable: true
          maxLength: 255
        is_primary:
          type: boolean
        is_emergency:
          type: boolean
        notes:
          type: string
          nullable: true
          maxLength: 5000
...
    PatientContactCollectionResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PatientContact'
    PatientContactMutationResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          examples: [patient_contact_created]
        data:
          $ref: '#/components/schemas/PatientContact'
...
    PatientDocument:
      type: object
      required:
        - id
        - patient_id
        - title
        - file_name
        - mime_type
        - size_bytes
        - uploaded_at
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        title:
          type: string
        document_type:
          type: string
          nullable: true
        file_name:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
        uploaded_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UploadPatientDocumentRequest:
      type: object
      required: [document]
      properties:
        document:
          type: string
          format: binary
          description: Allowed file types are `pdf`, `jpg`, `jpeg`, `png`, and `webp` up to `10 MiB`.
        title:
          type: string
          nullable: true
          maxLength: 255
        document_type:
          type: string
          nullable: true
          maxLength: 64
    PatientDocumentResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/PatientDocument'
    PatientDocumentCollectionResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PatientDocument'
    PatientDocumentMutationResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          examples: [patient_document_uploaded]
        data:
          $ref: '#/components/schemas/PatientDocument'
...
    ApiErrorResponse:
      type: object
      required: [code, message, details, trace_id, correlation_id]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        trace_id:
          type: string
          format: uuid
        correlation_id:
          type: string
          format: uuid